Further refactoring of collections

+++ src/Serval.Translation/Contracts/PhraseDto.cs
@@ -5,5 +5,5 @@ public class PhraseDto
     public int SourceSegmentStart { get; set; }
     public int SourceSegmentEnd { get; set; }
     public int TargetSegmentCut { get; set; }
-    public double Confidence { get; set; }
+    public float Confidence { get; set; }
 }
